zb 修改

+++ frameworks/util/network/src/stream_buffer.cpp
@@ -42,7 +42,7 @@ void StreamBuffer::Clean()
     }
 }
 
-bool StreamBuffer::SetReadIdx(size_t idx)
+bool StreamBuffer::SetReadIdx(int32_t idx)
 {
     if (idx > wIdx_) {
         MMI_LOGE("Invalid parameter input");
@@ -60,7 +60,7 @@ bool StreamBuffer::Read(std::string &buf)
         return false;
     }
     buf = ReadBuf();
-    rIdx_ += buf.size() + 1;
+    rIdx_ += static_cast<int32_t>(buf.size()) + 1;
     return (buf.size() > 0);
 }
 
@@ -94,7 +94,7 @@ bool StreamBuffer::Read(char *buf, size_t size)
         rwErrorStatus_ = ErrorStatus::ERROR_STATUS_READ;
         return false;
     }
-    if (rIdx_ + size > wIdx_) {
+    if (rIdx_ + static_cast<int32_t>(size) > wIdx_) {
         MMI_LOGE("Memory out of bounds on read... errCode:%{public}d", MEM_OUT_OF_BOUNDS);
         rwErrorStatus_ = ErrorStatus::ERROR_STATUS_READ;
         return false;
@@ -104,7 +104,7 @@ bool StreamBuffer::Read(char *buf, size_t size)
         rwErrorStatus_ = ErrorStatus::ERROR_STATUS_READ;
         return false;
     }
-    rIdx_ += size;
+    rIdx_ += static_cast<int32_t>(size);
     rCount_ += 1;
     return true;
 }
@@ -124,17 +124,17 @@ bool StreamBuffer::Write(const char *buf, size_t size)
         rwErrorStatus_ = ErrorStatus::ERROR_STATUS_WRITE;
         return false;
     }
-    if (wIdx_ + size >= MAX_STREAM_BUF_SIZE) {
+    if (wIdx_ + static_cast<int32_t>(size) >= MAX_STREAM_BUF_SIZE) {
         MMI_LOGE("The write length exceeds buffer. errCode:%{public}d", MEM_OUT_OF_BOUNDS);
         rwErrorStatus_ = ErrorStatus::ERROR_STATUS_WRITE;
         return false;
     }
-    if (EOK != memcpy_sp(&szBuff_[wIdx_], (MAX_STREAM_BUF_SIZE - wIdx_), buf, size)) {
+    if (EOK != memcpy_sp(&szBuff_[wIdx_], static_cast<size_t>(MAX_STREAM_BUF_SIZE - wIdx_), buf, size)) {
         MMI_LOGE("memcpy_sp call fail. errCode:%{public}d", MEMCPY_SEC_FUN_FAIL);
         rwErrorStatus_ = ErrorStatus::ERROR_STATUS_WRITE;
         return false;
     }
-    wIdx_ += size;
+    wIdx_ += static_cast<int32_t>(size);
     wCount_ += 1;
     return true;
 }
@@ -149,7 +149,7 @@ bool StreamBuffer::IsEmpty()
 
 size_t StreamBuffer::Size() const
 {
-    return wIdx_;
+    return static_cast<size_t>(wIdx_);
 }
 
 size_t StreamBuffer::UnreadSize() const
@@ -158,7 +158,7 @@ size_t StreamBuffer::UnreadSize() const
         MMI_LOGW("Widx_ less than ridx_, wIdx_:%{public}d,rIdx_:%{public}d", wIdx_, rIdx_);
         return 0;
     }
-    return (wIdx_ - rIdx_);
+    return static_cast<size_t>(wIdx_ - rIdx_);
 }
 
 bool StreamBuffer::ChkRWError() const
